Initial attempt to add simple binary ops to the SCCL interpreter
For now, to provide minimal testable functionality, this just performs
the same binary op as is performed by the "reduce" instructions (i.e.,
it's easy to use this to perform addition but not multiplication).
I'm not sure what the plan is for how to specify the type of the
binary op/reduction function to the interpreter but one would think
there should be a consistent way to do this, unless we're really
confident the only type of reduction we'll care about are sums.
Wouldn't we want to support other types of reduction ops like min and
max?