AllToAll should do reductions of the type indicated by the kernel launch, not always int8_t
In NVIDIA's version of NCCL, AllToAll always operates on raw bytes: it
can ignore the type of the data.  Not so for the SCCL interpreter, which
needs to know the datatype and can perform communication operations that
involve reductions (e.g.  RecvReduceSend, RecvReduceCopy).

Basically, I'm just going into all the places where AllToAll is treated
specially in terms of its datatype and making it look more like
AllReduce.

There's also one backwards-compatibility-breaking change here: ArgsCheck
(in argcheck.cc) no longer automatically scales up the buffer size by
the number of ranks, so there's no longer a need for applications that
call AllToAll to divide the buffer size they pass in by the number of ranks.